+++ equ-clt/src/main/java/equ/clt/handler/astron/AstronHandler.java
@@ -36,9 +36,9 @@ public class AstronHandler extends DefaultCashRegisterHandler<AstronSalesBatch> 
     }
 
     @Override
-    public Map<Integer, SendTransactionBatch> sendTransaction(List<TransactionCashRegisterInfo> transactionList) throws IOException {
+    public Map<Long, SendTransactionBatch> sendTransaction(List<TransactionCashRegisterInfo> transactionList) throws IOException {
 
-        Map<Integer, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
+        Map<Long, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
 
         if (transactionList != null) {
 
@@ -98,7 +98,7 @@ public class AstronHandler extends DefaultCashRegisterHandler<AstronSalesBatch> 
                             exportFlags(conn, "'GRP', 'ART', 'UNIT', 'PACK', 'EXBARC', 'PACKPRC'");
                             Throwable waitResult = waitFlags(conn, timeout);
                             if (waitResult != null) {
-                                for (Integer id : sendTransactionBatchMap.keySet()) {
+                                for (Long id : sendTransactionBatchMap.keySet()) {
                                     sendTransactionBatchMap.put(id, new SendTransactionBatch(waitResult));
                                 }
                             }
@@ -573,7 +573,7 @@ public class AstronHandler extends DefaultCashRegisterHandler<AstronSalesBatch> 
 
     @Override
     public void requestSalesInfo(List<RequestExchange> requestExchangeList, Set<String> directorySet,
-                                 Set<Integer> succeededRequests, Map<Integer, Throwable> failedRequests, Map<Integer, Throwable> ignoredRequests) throws IOException, ParseException {
+                                 Set<Long> succeededRequests, Map<Long, Throwable> failedRequests, Map<Long, Throwable> ignoredRequests) throws IOException, ParseException {
         AstronSettings astronSettings = springContext.containsBean("astronSettings") ? (AstronSettings) springContext.getBean("astronSettings") : null;
         String connectionString = astronSettings == null ? null : astronSettings.getConnectionString(); //"jdbc:mysql://172.16.0.35/export_axapta"
         String user = astronSettings == null ? null : astronSettings.getUser(); //luxsoft
