+++ equ-clt/src/main/java/equ/clt/handler/cas/CL5000JHandler.java
@@ -35,9 +35,9 @@ public class CL5000JHandler extends ScalesHandler {
     }
 
     @Override
-    public Map<Integer, SendTransactionBatch> sendTransaction(List<TransactionScalesInfo> transactionList) throws IOException {
+    public Map<Long, SendTransactionBatch> sendTransaction(List<TransactionScalesInfo> transactionList) throws IOException {
 
-        Map<Integer, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
+        Map<Long, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
 
         if (transactionList.isEmpty()) {
             processTransactionLogger.error("CL5000: Empty transaction list!");
